Move the search query builder into the SDK
The q string that /search-api parses is built and re-parsed by a grammar that
has to agree with the API's own parser, token for token. Until now that grammar
lived in apps/web, so the mobile app could not reuse it. Giving mobile advanced
search meant either importing across apps or writing a third implementation of
a grammar that already has to stay in lockstep with two - which is the exact
divergence that produced most of the defects in ecency/vision-next#1343.

The module moves verbatim: 219 lines, no imports, no behaviour change. It fits
the SDK's lightweight-utilities scope precisely because it depends on nothing.
Its spec moves with it, 51 tests.

Web now imports the same symbols from @ecency/sdk. Six call sites, imports only.

Two spec mocks needed the real implementation rather than a stub: the global
@ecency/sdk factory in setup-any-spec.ts, and the local one in the side-panels
spec. Both parse a URL's q with SearchQuery, and a stub would make every
assertion about parsing meaningless.

Typecheck note: apps/web resolves @ecency/sdk to the committed dist, and CI
typechecks without building packages first, so this branch is red until the
patch label rebuilds dist. Verified locally that a `pnpm --filter @ecency/sdk
build` takes web's tsc to 0; the built output is deliberately not committed.
