+++ packages/sdk/src/modules/search/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
+export * from "./query-builder";
 export * from "./queries";
 export * from "./requests";
